+++ generators/typescript/sdk/client-class-generator/src/GeneratedSdkClientClassImpl.ts
@@ -405,11 +405,25 @@ export class GeneratedSdkClientClassImpl implements GeneratedSdkClientClass {
         }
 
         if (this.shouldGenerateAuthorizationHeaderHelperMethod()) {
+            const returnsMaybeAuth =
+                !this.intermediateRepresentation.sdkConfig.isAuthMandatory || this.basicAuthScheme != null;
+            const returnType = returnsMaybeAuth
+                ? ts.factory.createTypeReferenceNode("Promise", [
+                      ts.factory.createUnionTypeNode([
+                          ts.factory.createKeywordTypeNode(ts.SyntaxKind.StringKeyword),
+                          ts.factory.createKeywordTypeNode(ts.SyntaxKind.UndefinedKeyword)
+                      ])
+                  ])
+                : ts.factory.createTypeReferenceNode("Promise", [
+                      ts.factory.createKeywordTypeNode(ts.SyntaxKind.StringKeyword)
+                  ]);
+
             serviceClass.addMethod({
                 scope: Scope.Protected,
                 isAsync: true,
                 name: GeneratedSdkClientClassImpl.AUTHORIZATION_HEADER_HELPER_METHOD_NAME,
-                statements: this.getAuthorizationHeaderStatements(context).map(getTextOfTsNode)
+                statements: this.getAuthorizationHeaderStatements(context).map(getTextOfTsNode),
+                returnType: getTextOfTsNode(returnType)
             });
         }