+++ seed/ts-sdk/exhaustive/dev-dependencies/src/api/resources/endpoints/resources/union/client/Client.ts
@@ -60,7 +60,7 @@ export class Union {
         };
     }
 
-    protected async _getAuthorizationHeader() {
+    protected async _getAuthorizationHeader(): Promise<string | undefined> {
         const bearer = await core.Supplier.get(this._options.token);
         if (bearer != null) {
             return `Bearer ${bearer}`;