+++ seed/ts-sdk/multi-url-environment/src/api/resources/s3/client/Client.ts
@@ -83,7 +83,7 @@ export class S3 {
         }
     }
 
-    protected async _getAuthorizationHeader() {
+    protected async _getAuthorizationHeader(): Promise<string> {
         return `Bearer ${await core.Supplier.get(this._options.token)}`;
     }
 }