+++ seed/ts-sdk/trace/exhaustive/src/api/resources/v2/client/Client.ts
@@ -70,7 +70,7 @@ export class V2 {
         return (this._v3 ??= new V3(this._options));
     }
 
-    protected async _getAuthorizationHeader() {
+    protected async _getAuthorizationHeader(): Promise<string | undefined> {
         const bearer = await core.Supplier.get(this._options.token);
         if (bearer != null) {
             return `Bearer ${bearer}`;