+++ seed/ts-sdk/trace/no-zurg-no-throwing/src/api/resources/migration/client/Client.ts
@@ -64,7 +64,7 @@ export class Migration {
         };
     }
 
-    protected async _getAuthorizationHeader() {
+    protected async _getAuthorizationHeader(): Promise<string | undefined> {
         const bearer = await core.Supplier.get(this._options.token);
         if (bearer != null) {
             return `Bearer ${bearer}`;