+++ seed/ts-sdk/trace/no-zurg-no-throwing/src/api/resources/playlist/client/Client.ts
@@ -324,7 +324,7 @@ export class Playlist {
         };
     }
 
-    protected async _getAuthorizationHeader() {
+    protected async _getAuthorizationHeader(): Promise<string | undefined> {
         const bearer = await core.Supplier.get(this._options.token);
         if (bearer != null) {
             return `Bearer ${bearer}`;