+++ seed/ts-sdk/trace/no-zurg-trace/src/api/resources/submission/client/Client.ts
@@ -249,7 +249,7 @@ export class Submission {
         }
     }
 
-    protected async _getAuthorizationHeader() {
+    protected async _getAuthorizationHeader(): Promise<string | undefined> {
         const bearer = await core.Supplier.get(this._options.token);
         if (bearer != null) {
             return `Bearer ${bearer}`;