+++ generators/typescript/sdk/CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.14.1-rc1] - 2024-04-12
+## [0.14.1-rc2] - 2024-04-17
+
+- Improvement: In an effort to make the generated code JSR compatible, we make sure all methods
+  are strongly typed with return signatures (in this case `_getAuthorizationHeader()`). 
+
+## [0.14.1-rc1] - 2024-04-17
 
 - Fix: Generate code snippet for FileDownload endpoint